+++ Source/Lib/Decoder/Codec/EbDecBitstream.h
@@ -78,7 +78,7 @@ void dec_bits_init(bitstrm_t *bs, const uint8_t *data, uint32_t u4_numbytes);
 
 uint32_t dec_get_bits_uvlc(bitstrm_t *bs);
 uint32_t dec_get_bits(bitstrm_t *bs, uint32_t numbits);
-void dec_get_bits_leb128(bitstrm_t *bs, size_t available, uint64_t *value,
+void dec_get_bits_leb128(bitstrm_t *bs, size_t available, size_t *value,
                     size_t *length);
 uint32_t dec_get_bits_ns(bitstrm_t *bs, uint32_t n);
 int32_t dec_get_bits_su(bitstrm_t *bs, uint32_t n);
